backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
langchain==0.0.350
langgraph==0.0.62
faiss-cpu==1.7.4
openai==1.3.7
python-socketio==5.10.0
redis==5.0.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1
google-generativeai==0.3.2